feat: single responsibility for layout, frontend in general
