+++ src/core/__init__.py
@@ -11,11 +11,11 @@ from .constants import *
 from .exceptions import *
 from .get_logger import get_logger
 from .structs import Addr
-from .util import Immutable
+from .util import Immutable, uninterruptible, process_redis_url
 
 __all__ = ["Addr", "StageEnum", "Immutable",
            "RCError", "AssignmentError", "NetworkError",
            "PartialResponseError", "PartialRequestError", "ConnectionCountError",
            "IS_CLI", "STAGE", "TLS_ENFORCED", "MAX_CONNECTIONS",
            "FILE_HANDLER", "STDOUT_HANDLER", "STDERR_HANDLER",
-           "get_logger"]
+           "get_logger", "uninterruptible", "process_redis_url"]
