+++ src/frontend/components/modals/manual_connect.py
@@ -8,7 +8,7 @@ class ManualConnect(ModalBase):
     A view for manually inputting Redis connection details (host, port, auth).
     """
 
-    def __init__(self, on_continue: Callable, switch_btn: ft.Control, close_btn: ft.Control):
+    def __init__(self, on_continue: Callable, switch_btn: ft.Button, close_btn: ft.IconButton):
         """
         Initialize the ManualConnect view with input fields.
