+++ src/main.py
@@ -6,25 +6,16 @@ from frontend import Layout
 
 from reactor import Reactor, ReactorClient
 from multiplexing import loop_multiplexing
-from util import uninterruptible
+from util import close_page, add_conn
+
 
 logger = core.get_logger(__name__)
 
-@uninterruptible
-async def close_page(multiplexing_event: Event, page: ft.Page) -> None:
-    logger.info("Closing application...")
-    try:
-        multiplexing_event.clear()
-    except BaseException as e:
-        logger.error(f"Application failed: {e}.", exc_info=True)
-    finally:
-        page.window.prevent_close = False
-        page.window.on_event = None
-        await page.window.destroy()
-        await page.window.close()
-
-@uninterruptible
+@core.uninterruptible
 def build_page(page: ft.Page) -> None:
+    """
+    Builds the main page.
+    """
     try:
         reactor = Reactor(page)
         multiplexing_event = Event()
@@ -34,10 +25,9 @@ def build_page(page: ft.Page) -> None:
         page.run_thread(loop_multiplexing, multiplexing_event, reactor)
         logger.info("Multiplexing thread started.")
         
-        async def handle_close(event: ft.WindowEvent | None = None) -> None:
-            if event and event.data != "close" and event.type != ft.WindowEventType.CLOSE:
-                return
-            await close_page(multiplexing_event, page)
+        async def handle_close(event: ft.WindowEvent | None) -> None:
+            if event is None or event.data == "close" or event.type == ft.WindowEventType.CLOSE:
+               await close_page(multiplexing_event, page)
         
         page.window.on_event = handle_close
         page.window.prevent_close = True
@@ -47,19 +37,18 @@ def build_page(page: ft.Page) -> None:
         # Handles OS intrusions gracefully.
         # Similar to a regular container.
         reactor_client = ReactorClient(reactor)
-        layout = Layout(reactor_client)
+        layout = Layout(lambda conn_data: add_conn(conn_data, reactor_client, layout))
         safe_area = ft.SafeArea(layout, expand=True)
         page.add(safe_area)
         logger.info("Flet app window initialized.")
     
     except Exception as e:
         logger.error(f"Application failed: {e}.", exc_info=True)
-        page.window.close()
+        page.run_task(page.window.close)
 
 if __name__ == "__main__":
     if core.IS_CLI:
         raise NotImplementedError("CLI mode is not implemented yet")
-        logger.info("CLI mode enabled.")
     else:
         logger.info("GUI mode enabled.")
         ft.run(build_page, assets_dir="src/frontend/assets")
